feat - #166 userResponse에서 여러 response를 관리할 수 있도록 변경
