+++ src/main/java/sopt/org/hmh/domain/user/controller/UserApi.java
@@ -3,8 +3,12 @@ package sopt.org.hmh.domain.user.controller;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestBody;
+import sopt.org.hmh.domain.user.dto.request.UserRequest.LockCheckDateRequest;
 import sopt.org.hmh.domain.user.dto.request.UserRequest.LockDateRequest;
-import sopt.org.hmh.domain.user.dto.response.UserInfoResponse;
+import sopt.org.hmh.domain.user.dto.response.UserResponse.IsLockTodayResponse;
+import sopt.org.hmh.domain.user.dto.response.UserResponse.UserInfoResponse;
+import sopt.org.hmh.global.auth.UserId;
 import sopt.org.hmh.global.common.response.BaseResponse;
 import sopt.org.hmh.global.common.response.EmptyJsonResponse;
 
@@ -19,13 +23,15 @@ public interface UserApi {
     @Operation(summary = "유저 포인트 정보 불러오기")
     ResponseEntity<BaseResponse<Integer>> orderGetUserPoint(@Parameter(hidden = true) final Long userId);
 
-    @Operation(
-            summary = "회원 탈퇴")
+    @Operation(summary = "회원 탈퇴")
     ResponseEntity<BaseResponse<EmptyJsonResponse>> orderWithdraw(@Parameter(hidden = true) final Long userId);
 
-    @Operation(
-            summary = "당일 잠금 여부 전송")
+    @Operation(summary = "당일 잠금 여부 전송")
     ResponseEntity<BaseResponse<EmptyJsonResponse>> orderChangeRecentLockDate(
             @Parameter(hidden = true) final Long userId, final LockDateRequest request);
+
+    @Operation(summary = "당일 잠금 여부 확인")
+    ResponseEntity<BaseResponse<IsLockTodayResponse>> orderGetRecentLockDate(
+            @UserId final Long userId, @RequestBody final LockCheckDateRequest request);
 
 }
